Add Japanese Hiragana and Katakana kana layouts
Two kana layouts (ja_JP, ja_JP_katakana) extracted 1:1 from the decompiled
reference, reusing the CombineMiddleware from the Hindi PR for dakuten voicing.

- Hiragana centers く つ い / ふ る ら / と ん す; Katakana ク ツ イ / …
- Kana filtered by Unicode block, dropping MessagEase's generic Latin ring.
- Dakuten voicing via combine: the dakuten key is remapped from MessagEase's
  internal quote to the real mark ゛(U+309B), and the combine rules are keyed
  by it (か + ゛ → が, は → ば, ワ → ヷ, …).
- Western digits (Japanese uses Arabic numerals); back-to-alpha label かな / カナ.

The growing MessagEase-layout extension is moved to its own file
LanguageDefinitions+MessagEase.swift to stay within SwiftLint's file_length.

Known limitations, matching MessagEase's base kana grid: handakuten (ぱ) and
small kana (ゃっ) are not reachable, and there is no kana→kanji conversion.

Tests: kana dakuten combine-rule assertions and an end-to-end UI test
(swipe か then ゛ → が). Stacked on #241/#243/#244/#247/#248.
